# api-spec.yml
openapi: 3.0.0
info:
  version: 0.1.0
  title: go-librespot API
  description: go-librespot daemon API
components:
  schemas:
...
    play:
      description: An initiate playback payload
      type: object
      properties:
        uri:
          description: Spotify URI to start playing
          type: string
        skip_to_uri:
          description: Spotify URI to skip to (when playing playlists)
          type: string
        paused:
          description: Start playback as paused
          type: boolean
        position:
          description: >-
            Position in milliseconds to start playback at within the selected
            track (0 starts from the beginning)
          type: integer
    cacheDownload:
      description: A pre-cache download payload
      type: object
      properties:
        uri:
          description: >-
            Spotify URI of the context (playlist, album, artist, track or
            episode) to download into the cache
          type: string
paths:
...
  /set_device_name:
    post:
      description: Set the player device name
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [ name ]
              properties:
                name:
                  description: The new device name
                  type: string
      responses:
        200:
          description: Successful response
  /player/play:
    post:
      description: >-
        Starts playing new content. To start a context shuffled (on a random
        track), enable shuffle first via /player/shuffle_context and then call
        this. Calling /player/shuffle_context after play only shuffles the
        upcoming queue and keeps the current (first) track, so play alone (or
        shuffle after play) always begins on the context's first track.
      requestBody:
        $ref: '#/components/schemas/play'
      responses:
        200:
          description: Successful response
  /cache/download:
    post:
      description: >-
        Download every track of a context (playlist, album, artist, track or
        episode) into the on-disk cache without playing it, so it can later be
        played without re-downloading from the CDN. Requires the cache to be
        enabled. Runs in the background and returns immediately; the download is
        paced to avoid hammering Spotify. Only the encrypted audio is stored, so
        the audio key is still required at playback time.
      requestBody:
        $ref: '#/components/schemas/cacheDownload'
      responses:
        200:
          description: Successful response
  /cache/snapshot:
    get:
      description: >-
        Return the current snapshot id (playlist revision) and track count for a
        playlist URI. Clients can compare the snapshot id against a previously
        stored value to decide whether the playlist changed and needs
        re-caching. Only playlists have a snapshot; for any other URI the
        snapshot_id is null.
      parameters:
        - name: uri
          in: query
          required: true
          description: Spotify playlist URI
          schema:
            type: string
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  snapshot_id:
                    type: string
                    nullable: true
                    description: The playlist snapshot id, or null for non-playlists
                  length:
                    type: integer
                    nullable: true
                    description: Number of tracks in the playlist
...
  /player/output:
    post:
      description: >-
        Reopen the audio output on a different device without touching the
        Spotify session, preserving the current track, playback position,
        paused state and volume. Useful for switching speakers (e.g. ALSA or
        bluealsa) live without reconnecting. If playback is active it switches
        devices immediately (with a brief audio gap); if nothing is playing the
        device is recorded and takes effect on the next output open. An empty
        device selects the configured default output device.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                device:
                  description: >-
                    The output device name to open (backend-specific, e.g. an
                    ALSA device). Empty selects the configured default.
                  type: string
      responses:
        200:
          description: Successful response
...
  /player/next:
    post:
      description: Skip to next track
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                uri:
                  type: string
                  description: The track URI to skip to
      responses:
        200:
          description: Successful response
...
  /player/seek:
    post:
      description: Seek current track
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [ position ]
              properties:
                position:
                  description: Seek position in milliseconds
                  type: integer
                relative:
                  description: Whether the seek position is relative to the current one
                  type: boolean
                  default: false
      responses:
        200:
          description: Successful response
  /player/volume:
    get:
      description: Get player volume
      responses:
        200:
          description: Returns the player volume settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  value:
                    type: integer
                    description: The current volume, ranging from 0 to max
                  max:
                    type: integer
                    description: The max volume value
    post:
      description: Set player volume
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [ volume ]
              properties:
                volume:
                  description: Volume from 0 to max
                  type: number
                relative:
                  description: Whether to change the volume relative to the current volume
                  type: boolean
                  default: false
      responses:
        200:
          description: Successful response
  /player/repeat_context:
    post:
      description: Toggle repeating context
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [ repeat_context ]
              properties:
                repeat_context:
                  description: Whether repeating context should be enabled
                  type: boolean
      responses:
        200:
          description: Successful response
  /player/repeat_track:
    post:
      description: Toggle repeating track
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [ repeat_track ]
              properties:
                repeat_track:
                  description: Whether repeating track should be enabled
                  type: boolean
      responses:
        200:
          description: Successful response
  /player/shuffle_context:
    post:
      description: >-
        Toggle shuffling the context. Enabling it BEFORE /player/play makes
        playback start on a random track; enabling it after play only reshuffles
        the upcoming queue and keeps the current (first) track.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [ shuffle_context ]
              properties:
                shuffle_context:
                  description: Whether shuffling context should be enabled
                  type: boolean
      responses:
        200:
          description: Successful response
  /player/add_to_queue:
    post:
      description: Add a track to the queue
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [ uri ]
              properties:
                uri:
                  description: The URI for the track that should be added
                  type: string
      responses:
        200:
          description: Successful response